further refactor - splitting things into classes.
